update to 1.0.5
use gitHub archive for source

+++ recipe/build.sh
@@ -6,6 +6,8 @@ ninja install
 # stuff gets installed into lib64 when we only use lib for conda
 if [[ ${HOST} =~ .*linux.* ]]; then
     mkdir -p $PREFIX/lib
-    mv $PREFIX/lib64/* $PREFIX/lib
-    rm -rf $PREFIX/lib64
+    if [[ -d $PREFIX/lib64 ]]; then
+        mv -f $PREFIX/lib64/* $PREFIX/lib
+        rm -rf $PREFIX/lib64
+    fi
 fi
